Added Backgrounds and Bring To Front toggles
Widget backgrounds can be toggled in the MCM - default is Off.  When enabled they become available in Edit Mode to show/hide/manipulate.  When disabled they are hidden and completely ignored in Edit Mode.

Bring To Front is now disabled by default but can be enabled in the MCM. Causes delay when switching presets and entering/leaving Edit Mode