Beta testing bug fixes
Throttle added to OnItemAdded which should protect against multiple items being added in rapid succession.
jArray.unique() call added to PotionScript sortPotionQueue as a failsafe to the above regarding the potion groups.
Moved a couple of the UI.Get calls in addToQueue to try and limit the possibility of wrong details being picked up if selected item focus changes once the function is running