Change to how we update item names on equip
Using WornObject.GetDisplayName() now as OnObjectEquipped doesn't receive akReference for very much at all so akReference.GetDisplayName() was returning empty strings for everything!

Made WC.findInQueue() a bit more longwinded but it works now with the form check if a form is passed, otherwise searches for a match by name.  Done this way so if we've just updated a name on equip it can find a match by form this time round.