Yeah, yeah. More 1.5 stuff
Added some additional code to the .as to check if the target clip is already displaying the desired frame and ignore as required

Adjusted the order calls are made in AddedItemHandler OnUpdate so adding back cached items no longer interferes with auto-equipping

Added option to exclude detrimental multi-effect potions from the potion groups

Allow ALL equipped items to be displayed in the widget regardless of auto-add options.  If auto-add isn't enabled for the option it will be flagged as a TempItem and removed from the queue again when OnObjectUnequipped runs.