More last minute bug fixing
Added poison dose to dummy Throwing Poison weapon on equipping to prevent player from being able to poison the weapon through the inventory menu.  Also displays the weapon as poisoned in the inventory list so seemed appropriate

Blocked CheckAndUpdatePoisonInfo if the target item is a throwing poison

Fix for Throwing Weapons messagebox showing up in game.